requirements.txt
# 核心依赖
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
PyYAML>=5.4.0

# 树模型
xgboost>=1.7.0
lightgbm>=3.3.0

# 特征编码
category_encoders>=2.5.0

# 可视化和评估
matplotlib>=3.5.0
seaborn>=0.11.0

# 节假日数据
chinese-calendar>=1.8.0
holidays>=0.25